requirements.txt
asgiref==3.2.10
astroid==2.4.2
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.3
Django==3.0.8
future==0.18.2
gunicorn==20.0.4
idna==2.10
isort==4.3.21
lazy-object-proxy==1.4.3
line-bot-sdk==1.16.0
mccabe==0.6.1
pylint==2.5.3
pytz==2020.1
requests==2.24.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
urllib3==1.25.9
wrapt==1.12.1
psycopg2==2.8.2
pillow==7.2.0
whitenoise==5.2.0
django-ckeditor==6.0.0
json5==0.9.5
djangorestframework==3.13.1